add first submodule POC
