+++ .gitignore
@@ -35,3 +35,7 @@ pip-delete-this-directory.txt
 .cache
 nosetests.xml
 coverage.xml
+
+.idea/
+venv/
+
